Bugfix R processing

+++ mods/fseq.sh
@@ -7,7 +7,7 @@
 # QCVARIABLES,Resource temporarily unavailable
 # RESULTFILENAME <DIR>/<TASK>/<SAMPLE>.bw
 
-echo ">>>>> DNase-Seq with fseq"
+echo ">>>>> Peak calling with fseq"
 echo ">>>>> startdate "`date`
 echo ">>>>> hostname "`hostname`
 echo ">>>>> job_name "$JOB_NAME
@@ -159,16 +159,17 @@ else
     cat > $OUTDIR/$SAMPLE.R <<DELIM    
 library(MASS)
 x <- read.delim("$OUTDIR/$SAMPLE.narrowPeak", header=F)
-g<-MASS::fitdistr(x$V7,"gamma")
-x$V8<-sapply(x$V7, function(y) ks.test(y, "pgamma",shape=g$estimate["shape"],rate=g$estimate["rate"])$p.value)
+g<-MASS::fitdistr(x\$V7,"gamma")
+x\$V8<-sapply(x\$V7, function(y) ks.test(y, "pgamma",shape=g\$estimate["shape"],rate=g\$estimate["rate"])\$p.value)
 write.table(x, "$OUTDIR/$SAMPLE.narrowPeak.tmp", sep="\t", row.names=FALSE, col.names=FALSE, quote=FALSE)
 DELIM
     
     Rscript --vanilla $OUTDIR/$SAMPLE.R
 
     #filter with p-value cutoff
-    awk -v CUTOFF=$FSEQ_PVALUECUTOFF '{if ($8 <= FSEQ_CUTOFF}{print $0}' $OUTDIR/$SAMPLE.narrowPeak.tmp > $OUTDIR/$SAMPLE.narrowPeak
-    rm $OUTDIR/$SAMPLE.narrowPeak.tmp
+    awk -v CUTOFF=$FSEQ_PVALUECUTOFF '{if ($8 <= CUTOFF){print $0}}' $OUTDIR/$SAMPLE.narrowPeak.tmp > $OUTDIR/$SAMPLE.narrowPeak
+    [ -f $OUTDIR/$SAMPLE.narrowPeak.tmp ] && rm $OUTDIR/$SAMPLE.narrowPeak.tmp 
+    [ -f $OUTDIR/$SAMPLE.R ] && rm $OUTDIR/$SAMPLE.R
 
     # mark checkpoint
     if [ -f $OUTDIR/$SAMPLE.narrowPeak ];then echo -e "\n********* $CHECKPOINT\n"; unset RECOVERFROM; else echo "[ERROR] checkpoint failed: $CHECKPOINT"; exit 1; fi
@@ -181,5 +182,5 @@ CHECKPOINT="cleanup"
 echo -e "\n********* $CHECKPOINT\n"
 ################################################################################
 [ -e $OUTDIR/$SAMPLE.bw.dummy ] && rm $OUTDIR/$SAMPLE.bw.dummy
-echo ">>>>> DNase-Seq with fseq - FINISHED"
+echo ">>>>> Peak calling with fseq - FINISHED"
 echo ">>>>> enddate "`date`
